pyproject.toml
[project]
name = "career-agent"
version = "0.1.0"
description = "AI Powered Career Assistant"
authors = [
    { name = "Yusuf Baykaloğlu", email = "yusuff.baykaloglu@gmail.com" }
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["agent", "ai", "job-search", "resume-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "streamlit>=1.45.1",
    "agno>=1.1.9",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=5.3.1",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "requests>=2.31.0",
    "pathlib>=1.0.1",
    "asyncio-throttle>=1.0.2",
]

[project.scripts]
career-agent = "app.streamlit_app:main"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
include = [
    "app/**/*.py",
    "README.md",
]
exclude = [
    "**/__pycache__",
    "**/*.pyc",
    "**/.*",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
